[connection] Made test mixins reusable (for extensions)
